requirements.txt
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
websocket-client==1.7.0
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.4.0
pykalman==0.9.7
python-dateutil==2.8.2
pytz==2023.3
colorlog==6.8.0
pytest==7.4.3
openpyxl==3.1.2